Re-design header component
