+++ package.json
@@ -5,10 +5,12 @@
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.27",
     "@fortawesome/free-solid-svg-icons": "^5.12.1",
+    "@popperjs/core": "^2.1.1",
     "bootstrap": "^4.4.1",
     "classnames": "^2.2.6",
     "date-fns": "^2.11.0",
     "firebase": "^5.11.1",
+    "node-sass": "^4.13.1",
     "rc-progress": "^2.5.2",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
@@ -17,6 +19,7 @@
     "react-responsive-modal": "^3.6.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
+    "reactstrap": "^8.4.1",
     "redux": "^4.0.5",
     "redux-firestore": "^0.7.4",
     "redux-thunk": "^2.3.0",
