+++ src/App.js
@@ -1,21 +1,22 @@
 import React, { Component } from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
-import Header from "./components/layout/Header";
+import Header from "./components/Layout/Header/Header";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Home from "./components/View/Home";
 import AddEntry from "./components/entries/AddEntry";
-import NotFound from "./components/layout/NotFound";
+import NotFound from "./components/Layout/NotFound";
 import store from "./store";
 import { Provider } from "react-redux";
+
+import './styles/share.scss';
 
 class App extends Component {
   render() {
     return (
       <Provider store={store}>
         <Router>
-          <div className="APP">
+          <div>
             <Header />
-
             <div className="container">
               <Switch>
                 <Route exact path="/" component={Home} />
