+++ src/components/ProgressBar/progress-bar.module.scss
@@ -1,5 +1,5 @@
 /* h3{
-    color:white;
+    color:white;.
     font-size: 24px
 } */
 
@@ -13,10 +13,7 @@
     align-content: center;
     align-items: center;
 }
-.flex {
-    display: flex;
-    flex-wrap: wrap;
-  }
+
 
   .head-right-container{
     justify-content: space-between;
