+++ src/components/layout/Header/header.module.scss
@@ -1,0 +1,43 @@
+.container {
+    max-width: inherit;
+}
+
+.navbar {
+    padding:0;
+
+}
+.navbar--green-color{
+    background-color:  rgb(0, 170, 70);
+}
+.navbar__content--site-name{
+    color:white;
+    font-size: 30px;
+    font-weight: bold;
+    display: inline-block;
+    padding-top: .3125rem;
+    padding-bottom: .3125rem;
+    margin-right: 1rem;
+    text-decoration: none;
+}
+.navbar__content--site-name:hover {  
+    text-decoration: none;
+}
+
+
+.navbar__main-right--list{
+    margin-bottom: 0;
+    list-style: none;
+}
+ 
+.navbar__main-right--icon{
+    color:white;
+    font-size: 23px;
+    text-decoration: none;
+    margin:5px
+}
+
+.navbar__main{
+    justify-content: space-between;
+    width: 100%;
+    align-items: center;
+}
