+++ src/styles/main.scss
@@ -1,0 +1,91 @@
+#prepend {
+    font-size: 16px;
+    min-width: 106px;
+    justify-content: center;
+  }
+  
+  #redButton {
+    display: flex;
+    position: relative;
+    line-height: 30px;
+    top: -50px;
+    float: right;
+  }
+  
+  .styles_modal__gNwvD {
+    max-width: 800px;
+    position: relative;
+    padding: 2.2rem;
+    background: #ffffff;
+    background-clip: padding-box;
+    box-shadow: 0 12px 15px 0 rgba(0, 0, 0, 0.25);
+    margin: auto;
+  }
+  
+  /* .form-control-lg {
+    height: calc(2.875rem + 2px);
+    padding: 0.5rem 7rem;
+    font-size: 1.25rem;
+    line-height: 1.5;
+    border-radius: 0.3rem;
+  } */
+  
+  #modelInput {
+    padding-left: 17px;
+    padding-top: 0;
+    padding-bottom: 0em;
+    width: 400px;
+    font-size: 17px;
+  }
+  
+  .invalid-feedback {
+    display: none;
+    margin-top: 0.25rem;
+    font-size: 80%;
+    color: #dc3545;
+  }
+  .form-control {
+    font-size: 1.25rem;
+  }
+  
+  .row-header {
+    margin-top: 0;
+    margin-bottom: -2rem;
+  }
+  .input-group > .custom-file,
+  .input-group > .custom-select,
+  .input-group > .form-control,
+  .input-group > .form-control-plaintext {
+    position: relative;
+    flex: 1 1 auto;
+  
+    margin-bottom: 0;
+  }
+  
+  .logo {
+    margin-right: 8px;
+    position: relative;
+    bottom: 1px;
+  }
+  
+  .navbar-brand {
+    font-weight: bold;
+  }
+  
+  .btn-light {
+    color: black;
+    background-color: #f8f9fa;
+    border-color: #f8f9fa;
+    font-weight: 500;
+  }
+  .bold-text {
+    font-weight: 500;
+  }
+  .bold-text-x {
+    font-weight: 700;
+  }
+  .blue-bg {
+    background-color: #2e36b5;
+    color: white;
+  }
+
