+++ src/styles/share.scss
@@ -1,3 +1,5 @@
+$primary-font-color: black;
+
 .bg-color{
     color:rgb(0, 170, 70);
 }
@@ -9,3 +11,7 @@
 .text-color{
     color:rgb(0, 170, 70);
 }
+.flex {
+    display: flex;
+    flex-wrap: wrap;
+  }
